🚧 修改用户模型和相关服务
✨ 将用户 ID 类型从 i64 更改为 u32

♻️ 重构用户服务和控制器以使用新的 SysUser 模型

🔧 更新数据库查询以使用新的字段名

🗃️ 调整主函数中的数据库查询

+++ src/models.rs
@@ -1132,7 +1132,7 @@ pub struct SysSignOn {
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct SysUser {
-    pub user_id: i64,
+    pub user_id: u32,
     pub user_name: String,
     pub nick_name: String,
     pub user_type: Option<i16>,
